rename: pet domain -> entity
