+++ src/test/java/animores/serverapi/to_do/service/impl/ToDoServiceImplTest.java
@@ -3,7 +3,7 @@ package animores.serverapi.to_do.service.impl;
 import animores.serverapi.account.domain.Account;
 import animores.serverapi.account.repository.AccountRepository;
 import animores.serverapi.common.exception.ExceptionCode;
-import animores.serverapi.pet.domain.Pet;
+import animores.serverapi.pet.entity.Pet;
 import animores.serverapi.pet.repository.PetRepository;
 import animores.serverapi.profile.domain.Profile;
 import animores.serverapi.profile.domain.vo.ProfileVo;
